+++ src/QNEthernetFrame.h
@@ -18,6 +18,7 @@
 #include <Stream.h>
 
 #include "StaticInit.h"
+#include "internal/PrintfChecked.h"
 #include "lwip/err.h"
 #include "lwip/netif.h"
 #include "lwip/pbuf.h"
@@ -37,7 +38,7 @@ namespace network {
 // 1. IPv4 (0x0800)
 // 2. ARP  (0x0806)
 // 3. IPv6 (0x86DD) (if enabled)
-class EthernetFrameClass final : public Stream {
+class EthernetFrameClass final : public Stream, public internal::PrintfChecked {
  public:
   // EthernetFrameClass is neither copyable nor movable
   EthernetFrameClass(const EthernetFrameClass &) = delete;
@@ -104,6 +105,9 @@ class EthernetFrameClass final : public Stream {
   //    or 18-(maxFrameLen()-4) for VLAN frames (excludes the FCS), or
   // 4. There's no room in the output buffers.
   bool send(const uint8_t *frame, size_t len) const;
+
+  // Use the one from here instead of the one from Print
+  using internal::PrintfChecked::printf;
 
   // Bring Print::write functions into scope
   using Print::write;